+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 
-## 4.18.0
+## 4.18.0+1
 April 01, 2024
 - New getters for App object: allowChangeTheme, allowChangeLocale, allowChangeUI, switchUI
 - if (widget.loadingScreen != null) { in app_statefulwidget.dart
 - Allows you to set an error handler more than once.
 - Removed deprecated errorHandler from runApp()
+- Commented out _materialColor() in app.dart
+- Updated show clause in run_app.dart
 
 ## 4.17.0
 February 09, 2023